Added 3 reminder emails Added a help page with tutorials with game, payment, mileage logging Changed sorting for a few pages Added an admin page with the ability to set a dashboard message as well as invite users Added a bunch of migrations to reflect database changes Added new routes to account for new pages / features

+++ app/Http/Controllers/MileageController.php
@@ -81,7 +81,7 @@ class MileageController extends Controller
             'date_travel' => $request->date_travel,
             'odometer_out' => $request->odometer_out,
             'odometer_in' => $request->odometer_in,
-            'distance' => is_null($request->odometer_in) && is_null($request->odometer_out) ? (int) $request->distance : (int) $request->odometer_in - (int) $request->odometer_out,
+            'distance' => is_null($request->odometer_in) && is_null($request->odometer_out) ? (double) $request->distance : (int) $request->odometer_in - (int) $request->odometer_out,
             'comments' => $request->comments,
             'status' => 'comp',
         ]);
